fix(ppu): rework audio-billing migration to not mutate merged #1124 history
The widen-to-Numeric migration had reused revision id f1a2b3c4d5e6 (already
merged on release-2.3 as the composite-index migration via #1124), renamed
that merged migration to a4b6c8d0e2f4, and edited merged d1e2f3a4b5c6's
down_revision. On any DB that had applied #1124 this desyncs alembic history
(widen ALTER skipped, composite index re-applied -> upgrade head fails).

- Restore f1a2b3c4d5e6_composite_index and d1e2f3a4b5c6 to their merged form
- Re-add the widen migration under a new unique id b7e2c9a4f1d3 chained on
  top of the current head f1a2b3c4d5e6
- Also widen ppu_tier_quotas.pending_monthly_quota (BigInteger -> Numeric(15,2))
  so the edit-tier staging/promotion path can hold fractional audio quotas;
  update the model and drop the now-unused BigInteger import

+++ services/platform-core-service/app/models/pay_per_use/ppu_tier.py
@@ -1,6 +1,6 @@
 import uuid
 
-from sqlalchemy import BigInteger, Boolean, Column, DateTime, ForeignKey, Numeric, String, Text, UniqueConstraint
+from sqlalchemy import Boolean, Column, DateTime, ForeignKey, Numeric, String, Text, UniqueConstraint
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
@@ -51,7 +51,7 @@ class PPUTierQuota(Base):
     )
     inference_name = Column(String(64), nullable=False)
     monthly_quota = Column(Numeric(15, 2), nullable=False)
-    pending_monthly_quota = Column(BigInteger, nullable=True)
+    pending_monthly_quota = Column(Numeric(15, 2), nullable=True)
     created_by = Column(String(255), nullable=True)
     updated_by = Column(String(255), nullable=True)
     created_at = Column(DateTime(timezone=True), nullable=False, server_default=func.now())
